requirements.txt
dj-static==0.0.6
Django==1.11.7
django-crispy-forms==1.7.0
pytz==2017.3
static3==0.7.0
